Changed Add Caption to PIL for unicode characters (#156)
* Changed Add Caption to PIL for unicode characters

* Added font, deleted assets, normalized image

* Fixed font path
